Redesign Auth page
