+++ lib/ui/page/auth/controller.dart
@@ -21,7 +21,6 @@ import 'package:get/get.dart';
 import 'package:rive/rive.dart';
 
 import '/domain/service/auth.dart';
-import '/l10n/_l10n.dart';
 import '/routes.dart';
 import '/util/message_popup.dart';
 
@@ -37,18 +36,11 @@ class AuthController extends GetxController {
   /// Current logo's animation frame.
   RxInt logoFrame = RxInt(0);
 
-  /// Index number of selected language.
-  late final RxInt selectedLanguage;
-
   /// A trigger of blink logo animation.
   SMITrigger? blink;
 
   /// [GlobalKey] of an animated button used to share it between overlays.
   final GlobalKey languageKey = GlobalKey();
-
-  /// Prevents instant language change after the user has set
-  ///  [selectedLanguage].
-  late final Worker _languageDebounce;
 
   /// Timer that periodically increases [logoFrame].
   Timer? _animationTimer;
@@ -57,25 +49,8 @@ class AuthController extends GetxController {
   Rx<RxStatus> get authStatus => _auth.status;
 
   @override
-  void onInit() {
-    selectedLanguage = RxInt(L10n.languages.keys.toList().indexOf(L10n.chosen));
-    _languageDebounce = debounce(
-      selectedLanguage,
-      (int i) {
-        L10n.chosen = L10n.languages.keys.elementAt(i);
-        Get.updateLocale(L10n.locales[L10n.chosen]!);
-      },
-      time: 500.milliseconds,
-    );
-
-    super.onInit();
-  }
-
-  @override
   void onClose() {
     _animationTimer?.cancel();
-    blink?.controller.dispose();
-    _languageDebounce.dispose();
     super.onClose();
   }
